Co-locate option types for hooks with the hooks
